+++ sdk/net/tls/session/quic_packet_publisher.cpp
@@ -58,7 +58,7 @@ quic_packet_publisher& quic_packet_publisher::set_flags(uint32 flags) {
 
 tls_session* quic_packet_publisher::get_session() { return _session; }
 
-uint16 quic_packet_publisher::get_payload_size() { return _payload_size; }
+size_t quic_packet_publisher::get_payload_size() { return _payload_size; }
 
 uint32 quic_packet_publisher::get_flags() { return _flags; }
